refactoring validate control logic
