--- node/src/reactor/main_reactor/validate_instruction.rs
@@ -1,12 +1,0 @@
-use std::time::Duration;
-
-use crate::{effect::Effects, reactor::main_reactor::MainEvent};
-
-pub(super) enum ValidateInstruction {
-    Do(Duration, Effects<MainEvent>),
-    CheckLater(String, Duration),
-    NonSwitchBlock,
-    KeepUp,
-    ShutdownForUpgrade,
-    Fatal(String),
-}
